Moved files, made corrections, and added message count functions
Moved files into the src folder and changed Logs to appear in there.

Made corrections to the bat and sh start scripts because of the moving of the files.

Added the message_history_count function to get the total count or a certain count of messages of a specified user.