Did a bunch of optimization, added Test Modal, and added additional logging
Optimized Log directory to be created correctly along with adding more logging in places as well as adding a datetime for each log.

Made sure when "testing" is disabled, all test commands are removed, and when enabled all test commands are added.

Added a Test Modal, seeing what can be done with that.

Changed check_admin to check for me and not the role Administrator.

Added a exception check in dm_test.

Made sure that only users are scanned in mod-help aren't other bots besides P2MM Bot.

Fixed up on_error function for error events.