Fixed error logging, fixed stuff for message_history_count, and mod-help message
Fixed error logging so it displays properly and doesn't error when displaying an error.

Fixed message_history_count with a read_limit variable so that it only reads the amount of messages it needs to find, that way the bot isn't rate limited everytime it tries to retrieve messages from the channel.

Fixed mod-help message to display as a reply message that disappears after 15 seconds instead of a ephemeral message because the interaction ephemeral message wasn't working without on_message supplying interaction context.